requirements.txt
pandas>=0.24.2
pythonping>=1.0.5
speedtest-cli>=2.1.2
jupyter>=1.0.0
holoviews>=1.12.5
bokeh>=1.3.4
seaborn>=0.9.0
schedule>=0.6.0
